Rename file names to GameCard/s
